requirements.txt
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
newspaper3k
textblob==0.18.0
Pillow==11.2.1
requests==2.32.3
lxml_html_clean==0.3.1



//  https://apnews.com/article/hurricane-melissa-farmers-jamaica-cuba-haiti-wfp-880f6182c6054f8b65a34872d9e211cc   //